// src/app/services/models/leave-balance-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EmployeeShortDto } from '../models/employee-short-dto';
export interface LeaveBalanceDto {
  adjustedBalance?: number;
  adjustmentReason?: string;
  currentBalance?: number;
  employee?: EmployeeShortDto;
  id?: number;
  initialBalance?: number;
  leaveType?: 'ANNUAL_LEAVE' | 'SICK_LEAVE' | 'MATERNITY_LEAVE' | 'PATERNITY_LEAVE' | 'UNPAID_LEAVE' | 'BEREAVEMENT_LEAVE' | 'MARRIAGE_LEAVE' | 'TRAINING_LEAVE' | 'SABBATICAL_LEAVE' | 'OTHER';
  usedBalance?: number;
  year?: number;
}