// src/app/services/models/leave-create-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface LeaveCreateDto {
  attachments?: string;
  employeeId?: number;
  endDate?: string;
  halfDay?: boolean;
  leaveType?: 'ANNUAL_LEAVE' | 'SICK_LEAVE' | 'MATERNITY_LEAVE' | 'PATERNITY_LEAVE' | 'UNPAID_LEAVE' | 'BEREAVEMENT_LEAVE' | 'MARRIAGE_LEAVE' | 'TRAINING_LEAVE' | 'SABBATICAL_LEAVE' | 'OTHER';
  reason?: string;
  startDate?: string;
}